requirements.txt
# Azure AI and services
azure-ai-agents==1.2.0b5
azure-ai-projects==1.1.0b4
azure-core==1.35.1
azure-identity==1.25.0

# OpenAPI and API validation
openapi-core==0.19.5

# Databricks
databricks_ai_bridge==0.8.0
databricks-sdk==0.67.0

# Bot Framework
botbuilder-core==4.17.0
botbuilder-integration-aiohttp==4.17.0
botbuilder-schema==4.17.0
botframework-connector==4.17.0
botframework-streaming==4.17.0

# FastAPI
fastapi
uvicorn[standard]
asyncio

# Production server
gunicorn==22.0.0

matplotlib